+++ R/utils.R
@@ -56,5 +56,28 @@ JS <- function(...) {
   x
 }
 
+list_remove_null <- function(x) {
+
+  # determine which elements are NULL
+  is_null <- vapply(x, is.null, logical(1))
+
+  # remove them by settiing them to NULL (!?!)
+  x[is_null] <- NULL
+
+  x
+}
+
+is_simple_chart <- function(x) {
+
+  class_simple_chart <-
+    c("altair.vegalite.v2.api.Chart", "altair.vegalite.v2.api.LayerChart")
+
+  inherits(x, class_simple_chart)
+}
+
+is_undefined <- function(x) {
+  inherits(x, "altair.utils.schemapi.UndefinedType")
+}
+