+++ docs/articles/first-example_files/vegalite-binding-0.1.9000/vegalite.js
@@ -10,12 +10,20 @@ HTMLWidgets.widget({
 
       renderValue: function(x) {
 
-        var spec = JSON.parse(x.spec);
+        var chart_spec = JSON.parse(x.chart_spec);
         var tooltip_options = x.tooltip_options;
+        var embed_options = x.embed_options;
 
-        vegaEmbed(el, spec).then(function(result) {
+        vegaEmbed(el, chart_spec, opt = embed_options).then(function(result) {
           // access view as result.view
-          vegaTooltip.vegaLite(result.view, spec, tooltip_options);
+          vegaTooltip.vegaLite(result.view, chart_spec, tooltip_options);
+
+          // By removing the style (width and height) of the
+          // enclosing element, we let the "chart" decide the space it
+          // will occupy.
+          //
+          el.removeAttribute("style");
+
         }).catch(console.error);
       },